+++ src/core/LabeledInput/LabeledInput.tsx
@@ -70,7 +70,7 @@ export const LabeledInput = React.forwardRef<
 
   useTheme();
 
-  const icon = status ? StatusIconMap[status] : svgIcon;
+  const icon = status ? StatusIconMap[status]() : svgIcon;
 
   return (
     <label